Extract file-related functions to trait
To be reused somewhere else.
